requirements.txt
torch~=1.8.1+cu102
torchvision~=0.9.1+cu102
numpy~=1.19.5
gym~=0.17.1
pandas~=1.0.3
opencv-python~=4.2.0.34
Pillow~=7.1.1
pygame~=1.9.6
stable-baselines3~=1.0
